kmmrce/requirements.txt
djangorestframework
django-cors-headers
pytz